requirements-vercel.txt
# Briefly Cloud - Production Optimized Dependencies
# Serverless deployment with essential OAuth functionality

# Core FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==24.1.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0

# Supabase client
supabase==2.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Stripe for payments
stripe==8.2.0

# Essential utilities
python-dateutil==2.8.2
requests==2.31.0

# Document parsing (lightweight)
python-docx==1.1.0

# External API clients
openai==1.3.0

# Google API libraries (required for OAuth)
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Microsoft Graph API
msal==1.24.1

# Deployment Mode: API-only (external ML services)
# Heavy ML libraries removed: chromadb, transformers, sentence-transformers
# Size reduction: ~6GB → ~500MB